test/requirements.txt
Pillow==9.2.0
requests==2.7.0
boto3==1.24.75
